cleaned up several more non-fixed sized integer types to be fixed for portability

+++ adios.c
@@ -85,11 +85,11 @@ void adios_allocate_buffer_ (int * err)
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_open (long long * fd, const char * group_name
+static int common_adios_open (int64_t * fd, const char * group_name
                              ,const char * name, const char * file_mode
                              )
 {
-    long long group_id = 0;
+    int64_t group_id = 0;
     struct adios_file_struct * fd_p = (struct adios_file_struct *)
                                   malloc (sizeof (struct adios_file_struct));
     struct adios_group_struct * g = 0;
@@ -153,19 +153,19 @@ static int common_adios_open (long long * fd, const char * group_name
         methods = methods->next;
     }
 
-    *fd = (long long) fd_p;
+    *fd = (int64_t) fd_p;
 
     return 0;
 }
 
-int adios_open (long long * fd, const char * group_name, const char * name
+int adios_open (int64_t * fd, const char * group_name, const char * name
                ,const char * mode
                )
 {
     return common_adios_open (fd, group_name, name, mode);
 }
 
-void adios_open_ (long long * fd, const char * group_name, const char * name
+void adios_open_ (int64_t * fd, const char * group_name, const char * name
                  ,const char * mode, int * err
                  ,int group_name_size, int name_size, int mode_size
                  )
@@ -186,7 +186,7 @@ void adios_open_ (long long * fd, const char * group_name, const char * name
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_group_size (long long fd_p
+static int common_adios_group_size (int64_t fd_p
                                    ,uint64_t data_size
                                    ,uint64_t * total_size
                                    ,void * comm
@@ -282,14 +282,14 @@ static int common_adios_group_size (long long fd_p
     return 0;
 }
 
-int adios_group_size (long long fd_p, uint64_t data_size
+int adios_group_size (int64_t fd_p, uint64_t data_size
                      ,uint64_t * total_size, void * comm
                      )
 {
     return common_adios_group_size (fd_p, data_size, total_size, comm);
 }
 
-void adios_group_size_ (long long * fd_p, int64_t * data_size
+void adios_group_size_ (int64_t * fd_p, int64_t * data_size
                        ,int64_t * total_size, void * comm, int * err
                        )
 {
@@ -299,7 +299,7 @@ void adios_group_size_ (long long * fd_p, int64_t * data_size
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_write (long long fd_p, const char * name, void * var)
+static int common_adios_write (int64_t fd_p, const char * name, void * var)
 {
     struct adios_file_struct * fd = (struct adios_file_struct *) fd_p;
     struct adios_var_struct * v = fd->group->vars;
@@ -413,12 +413,12 @@ static int common_adios_write (long long fd_p, const char * name, void * var)
     return 0;
 }
 
-int adios_write (long long fd_p, const char * name, void * var)
+int adios_write (int64_t fd_p, const char * name, void * var)
 {
     return common_adios_write (fd_p, name, var);
 }
 
-void adios_write_ (long long * fd_p, const char * name, void * var, int * err
+void adios_write_ (int64_t * fd_p, const char * name, void * var, int * err
                   ,int name_size
                   )
 {
@@ -433,7 +433,7 @@ void adios_write_ (long long * fd_p, const char * name, void * var, int * err
 
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_get_write_buffer (long long fd_p, const char * name
+static int common_adios_get_write_buffer (int64_t fd_p, const char * name
                                          ,uint64_t * size
                                          ,void ** buffer
                                          )
@@ -484,7 +484,7 @@ static int common_adios_get_write_buffer (long long fd_p, const char * name
     return 0;
 }
 
-int adios_get_write_buffer (long long fd_p, const char * name
+int adios_get_write_buffer (int64_t fd_p, const char * name
                            ,uint64_t * size
                            ,void ** buffer
                            )
@@ -492,8 +492,8 @@ int adios_get_write_buffer (long long fd_p, const char * name
     return common_adios_get_write_buffer (fd_p, name, size, buffer);
 }
 
-void adios_get_write_buffer_ (long long * fd_p, const char * name
-                             ,uint64_t * size
+void adios_get_write_buffer_ (int64_t * fd_p, const char * name
+                             ,int64_t * size
                              ,void ** buffer, int * err, int name_size
                              )
 {
@@ -507,7 +507,7 @@ void adios_get_write_buffer_ (long long * fd_p, const char * name
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_read (long long fd_p, const char * name, void * buffer
+static int common_adios_read (int64_t fd_p, const char * name, void * buffer
                              ,uint64_t buffer_size
                              )
 {
@@ -559,15 +559,15 @@ static int common_adios_read (long long fd_p, const char * name, void * buffer
     return 0;
 }
 
-int adios_read (long long fd_p, const char * name, void * buffer
+int adios_read (int64_t fd_p, const char * name, void * buffer
                ,uint64_t buffer_size
                )
 {
     return common_adios_read (fd_p, name, buffer, buffer_size);
 }
 
-void adios_read_ (long long * fd_p, const char * name, void * buffer
-                 ,long long * buffer_size, int * err, int name_size
+void adios_read_ (int64_t * fd_p, const char * name, void * buffer
+                 ,int64_t * buffer_size, int * err, int name_size
                  )
 {
     char * buf1 = 0;
@@ -580,7 +580,7 @@ void adios_read_ (long long * fd_p, const char * name, void * buffer
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_set_path (long long fd_p, const char * path)
+static int common_adios_set_path (int64_t fd_p, const char * path)
 {
     struct adios_file_struct * fd = (struct adios_file_struct *) fd_p;
     struct adios_group_struct * t = fd->group;
@@ -614,12 +614,12 @@ static int common_adios_set_path (long long fd_p, const char * path)
     return 0;
 }
 
-int adios_set_path (long long fd_p, const char * path)
+int adios_set_path (int64_t fd_p, const char * path)
 {
     return common_adios_set_path (fd_p, path);
 }
 
-void adios_set_path_ (long long * fd_p, const char * path, int * err
+void adios_set_path_ (int64_t * fd_p, const char * path, int * err
                      ,int path_size
                      )
 {
@@ -633,7 +633,7 @@ void adios_set_path_ (long long * fd_p, const char * path, int * err
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_set_path_var (long long fd_p, const char * path
+static int common_adios_set_path_var (int64_t fd_p, const char * path
                                      ,const char * name
                                      )
 {
@@ -665,12 +665,12 @@ static int common_adios_set_path_var (long long fd_p, const char * path
     return 0;
 }
 
-int adios_set_path_var (long long fd_p, const char * path, const char * name)
+int adios_set_path_var (int64_t fd_p, const char * path, const char * name)
 {
     return common_adios_set_path_var (fd_p, path, name);
 }
 
-void adios_set_path_var_ (long long * fd_p, const char * path
+void adios_set_path_var_ (int64_t * fd_p, const char * path
                          ,const char * name, int * err, int path_size
                          ,int name_size
                          )
@@ -784,7 +784,7 @@ void adios_stop_calculation_ (int * err)
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-static int common_adios_close (long long fd_p)
+static int common_adios_close (int64_t fd_p)
 {
     struct adios_file_struct * fd = (struct adios_file_struct *) fd_p;
     struct adios_method_list_struct * m = fd->group->methods;
@@ -859,12 +859,12 @@ static int common_adios_close (long long fd_p)
     return 0;
 }
 
-int adios_close (long long fd_p)
+int adios_close (int64_t fd_p)
 {
     return common_adios_close (fd_p);
 }
 
-void adios_close_ (long long * fd_p, int * err)
+void adios_close_ (int64_t * fd_p, int * err)
 {
     *err = common_adios_close (*fd_p);
 }
@@ -877,7 +877,7 @@ void adios_close_ (long long * fd_p, int * err)
 
 ///////////////////////////////////////////////////////////////////////////////
 // group a list of vars into a composite group
-int adios_declare_group (long long * id, const char * name
+int adios_declare_group (int64_t * id, const char * name
                         ,const char * coordination_comm
                         ,const char * coordination_var
                         ,const char * time_index
@@ -890,7 +890,7 @@ int adios_declare_group (long long * id, const char * name
                                       );
 }
 
-void adios_declare_group_ (long long * id, const char * name
+void adios_declare_group_ (int64_t * id, const char * name
                           ,const char * coordination_comm
                           ,const char * coordination_var
                           ,const char * time_index, int * err
@@ -923,7 +923,7 @@ void adios_declare_group_ (long long * id, const char * name
 // adios_common_define_var is in adios_internals.c
 
 // declare a single var as an entry in a group
-int adios_define_var (long long group_id, const char * name
+int adios_define_var (int64_t group_id, const char * name
                      ,const char * path, int type
                      ,const char * dimensions
                      ,const char * global_dimensions
@@ -938,7 +938,7 @@ int adios_define_var (long long group_id, const char * name
 }
 
 // declare a single var as an entry in a group
-void adios_define_var_ (long long * group_id, const char * name
+void adios_define_var_ (int64_t * group_id, const char * name
                        ,const char * path, int * type
                        ,const char * dimensions
                        ,const char * global_dimensions
@@ -975,7 +975,7 @@ void adios_define_var_ (long long * group_id, const char * name
 
 // adios_common_define_attribute is in adios_internals.c
 
-int adios_define_attribute (long long group, const char * name
+int adios_define_attribute (int64_t group, const char * name
                            ,const char * path, enum ADIOS_DATATYPES type
                            ,const char * value, const char * var
                            )
@@ -983,7 +983,7 @@ int adios_define_attribute (long long group, const char * name
     return adios_common_define_attribute (group, name, path, type, value, var);
 }
 
-void adios_define_attribute_ (long long * group, const char * name
+void adios_define_attribute_ (int64_t * group, const char * name
                              ,const char * path, int type, const char * value
                              ,const char * var, int * err
                              ,int name_size, int path_size, int value_size
